Move files due to migration to Maven project
