+++ SourceCode/LogParser/src/main/java/Interfaces/Decompressable.java
@@ -3,7 +3,6 @@ package Interfaces;
 import java.io.File;
 import java.util.List;
 
-public interface IDecompressable {
-
+public interface Decompressable {
     List<File> decompress(File archive);
 }
